pyproject.toml
[tool.poetry]
name = "agentlebot"
version = "0.1.0"
description = "A Gentle AI Bot to do your things for you, with memory and functions."
authors = ["Anshul Aggarwal <anshulaggarw4l@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
mypy = "^1.7.1"
pydantic = "^2.5.2"
pinecone-client = "^2.2.4"
openai = "^1.3.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
